add customer-share

+++ src/main/java/com/example/crm/controller/crm/LinkManController.java
@@ -1,6 +1,7 @@
 package com.example.crm.controller.crm;
 
 import com.example.crm.entity.Customer;
+import com.example.crm.entity.Employee;
 import com.example.crm.entity.LinkMan;
 import com.example.crm.repository.CustomerRepository;
 import com.example.crm.repository.LinkManRepository;
@@ -25,13 +26,20 @@ public class LinkManController {
 
 
     @GetMapping("")
-    public ModelAndView linkMen(@RequestParam Integer customerId){
+    public ModelAndView linkMen(@RequestParam Integer customerId, @SessionAttribute Employee loginEmployee){
         Optional<Customer> customerOptional = customerRepository.findById(customerId);
         ModelAndView mav = new ModelAndView("link_man");
         if(customerOptional.isPresent()){
             Customer customer = customerOptional.get();
+
+            // 没有查看该用户联系人的权限
+            if(!loginEmployee.getCustomers().contains(customer)){
+                return mav;
+            }
+
             List<LinkMan> linkMen = customer.getLinkMen();
             mav.addObject("linkMans", linkMen);
+            mav.addObject("customerName", customer.getName());
             mav.addObject("customerId", customerId);
             return mav;
         }
@@ -39,87 +47,107 @@ public class LinkManController {
     }
 
     @PostMapping(path = "")
-    public ResponseEntity<Map<String, String>> add(@RequestBody JSONObject jsonObject) {
+    public ResponseEntity<Map<String, String>> add(@RequestBody JSONObject jsonObject, @SessionAttribute Employee loginEmployee) {
+        Map<String, String> responseMap = new HashMap<>();
         if(!jsonDataCheck(jsonObject)){
-            Map<String, String> map = new HashMap<>();
-            map.put("message", "error data type");
-            return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+            responseMap.put("message", "数据错误");
+            return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
         }
 
         LinkMan linkMan = new LinkMan(jsonObject.getString("name"), jsonObject.getString("tel"));
         Optional<Customer> optionalCustomer = customerRepository.findById(jsonObject.getInt("customer_id"));
         if(optionalCustomer.isPresent()){
             Customer customer = optionalCustomer.get();
+
+            if(!loginEmployee.getCustomers().contains(customer)){
+                responseMap.put("message", "没有权限");
+                return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
+            }
+
             List<LinkMan> linkmen = customer.getLinkMen();
             linkmen.add(linkMan);
             customer.setLinkMen(linkmen);
             linkManRepository.save(linkMan);
             customerRepository.save(customer);
 
-            Map<String, String> map = new HashMap<>();
-            map.put("message", "success");
-            return new ResponseEntity<>(map, HttpStatus.OK);
+            responseMap.put("message", "添加成功");
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
         }
-        Map<String, String> map = new HashMap<>();
-        map.put("message", "linkman do not exist");
-        return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+        responseMap.put("message", "客户不存在");
+        return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
     }
 
     @PostMapping(path = "update")
-    public ResponseEntity<Map<String, String>> update(@RequestBody JSONObject jsonObject) {
+    public ResponseEntity<Map<String, String>> update(@RequestBody JSONObject jsonObject, @SessionAttribute Employee loginEmployee) {
+        Map<String, String> responseMap = new HashMap<>();
         if(!jsonDataCheck(jsonObject)){
-                Map<String, String> map = new HashMap<>();
-                map.put("message", "error data type");
-                return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+            responseMap.put("message", "数据错误");
+                return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
         }
 
         Optional<LinkMan> optionalLinkMan = linkManRepository.findById(jsonObject.getInt("id"));
-        if(optionalLinkMan.isPresent()){
+        Optional<Customer> optionalCustomer = customerRepository.findById(jsonObject.getInt("customer_id"));
+        if(optionalLinkMan.isPresent() && optionalCustomer.isPresent()){
+            Customer customer = optionalCustomer.get();
+
+            if(!loginEmployee.getCustomers().contains(customer)){
+                responseMap.put("message", "没有权限");
+                return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
+            }
+
             LinkMan lkm = optionalLinkMan.get();
             lkm.setName(jsonObject.getString("name"));
             lkm.setTel(jsonObject.getString("tel"));
             linkManRepository.save(lkm);
 
-            Map<String, String> map = new HashMap<>();
-            map.put("message", "success");
-            return new ResponseEntity<>(map, HttpStatus.OK);
+            responseMap.put("message", "修改成功");
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
         }
-        Map<String, String> map = new HashMap<>();
-        map.put("message", "linkman do not exist");
-        return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+        responseMap.put("message", "客户不存在");
+        return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
     }
 
     @PostMapping(path = "delete")
-    public ResponseEntity<Map<String, String>> delete(@RequestBody JSONObject jsonObject) {
+    public ResponseEntity<Map<String, String>> delete(@RequestBody JSONObject jsonObject, @SessionAttribute Employee loginEmployee) {
+        Map<String, String> responseMap = new HashMap<>();
         try{
             jsonObject.getInt("id");
+            jsonObject.getInt("customer_id");
         }
-        catch (JSONException e){
-            Map<String, String> map = new HashMap<>();
-            map.put("message", "data type error");
-            return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+        catch (Exception e){
+            responseMap.put("message", "数据错误");
+            return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
         }
 
         Optional<LinkMan> optionalLinkMan = linkManRepository.findById(jsonObject.getInt("id"));
-        if(optionalLinkMan.isPresent()){
+        Optional<Customer> optionalCustomer = customerRepository.findById(jsonObject.getInt("customer_id"));
+        if(optionalLinkMan.isPresent() && optionalCustomer.isPresent()){
+
+            Customer customer = optionalCustomer.get();
+            if(!loginEmployee.getCustomers().contains(customer)){
+                responseMap.put("message", "没有权限");
+                return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
+            }
+
             linkManRepository.delete(optionalLinkMan.get());
 
-            Map<String, String> map = new HashMap<>();
-            map.put("message", "success");
-            return new ResponseEntity<>(map, HttpStatus.OK);
+            responseMap.put("message", "删除成功");
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
         }
 
-        Map<String, String> map = new HashMap<>();
-        map.put("message", "linkman did not exist");
-        return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
+        responseMap.put("message", "联系人不存在");
+        return new ResponseEntity<>(responseMap, HttpStatus.BAD_REQUEST);
     }
 
     private boolean jsonDataCheck(JSONObject jsonObject){
         try{
             if(jsonObject.get("id")!=null && jsonObject.get("id")!="")
                 jsonObject.getInt("id");
+            jsonObject.getInt("customer_id");
             jsonObject.getString("name");
             jsonObject.getString("tel");
+            if(jsonObject.getString("name").equals("")||jsonObject.getString("tel").equals(""))
+                return false;
         }catch (Exception e){
             return false;
         }
